+++ orix/plot/inverse_pole_figure_plot.py
@@ -32,7 +32,6 @@ import matplotlib.projections as mprojections
 import matplotlib.pyplot as plt
 import numpy as np
 
-from orix.crystal_map.phase_list import Phase
 from orix.measure.pole_density_function import pole_density_function
 from orix.plot.direction_color_keys.direction_color_key_tsl import DirectionColorKeyTSL
 from orix.plot.stereographic_plot import ZORDER, StereographicPlot
@@ -475,10 +474,15 @@ def _get_ipf_axes_labels(vertices: Vector3d, symmetry: Symmetry) -> list[str]:
     list of str
         List of strings, with -1 formatted like $\bar{1}$.
     """
+    # Import here to avoid circular import
+    from orix.crystal_map._phase import Phase
+
     phase = Phase(point_group=symmetry)
-    m = Miller(uvw=vertices.data, phase=phase)
+    m = Miller(xyz=vertices.data, phase=phase)
     if symmetry.system in ["trigonal", "hexagonal"]:
         m.coordinate_format = "UVTW"
+    else:
+        m.coordinate_format = "uvw"
     axes = m.round(max_index=2).coordinates.astype(int)
 
     labels = []