+++ orix/quaternion/symmetry.py
@@ -19,7 +19,7 @@
 
 from __future__ import annotations
 
-from typing import TYPE_CHECKING
+from typing import TYPE_CHECKING, Literal
 
 from diffpy.structure.spacegroups import GetSpaceGroup
 import matplotlib.figure as mfigure
@@ -31,6 +31,17 @@ from orix.vector.vector3d import Vector3d
 if TYPE_CHECKING:  # pragma: no cover
     from orix.quaternion.orientation import Orientation
     from orix.vector.fundamental_sector import FundamentalSector
+
+
+VALID_SYSTEMS = Literal[
+    "triclinic",
+    "monoclinic",
+    "orthorhombic",
+    "tetragonal",
+    "trigonal",
+    "hexagonal",
+    "cubic",
+]
 
 
 class Symmetry(Rotation):
@@ -158,14 +169,9 @@ class Symmetry(Rotation):
         return region
 
     @property
-    def system(self) -> str | None:
+    def system(self) -> VALID_SYSTEMS | None:
         """Return which of the seven crystal systems this symmetry
-        belongs to.
-
-        Returns
-        -------
-        system
-            ``None`` is returned if the symmetry name is not recognized.
+        belongs to, or None if the symmetry name is not recognized.
         """
         name = self.name
         if name in ["1", "-1"]: