+++ orix/tests/test_crystal_map/test_crystal_map.py
@@ -970,9 +970,9 @@ class TestCrystalMapRepresentation:
         # fmt: off
         assert repr(xmap) == (
             "Phase  Orientations  Name  Space group  Point group  Proper point group  Color\n"
-            "    0    10 (83.3%)     a        Im-3m         m-3m                 432      r\n"
-            "    1      1 (8.3%)     b         P432          432                 432      g\n"
-            "    2      1 (8.3%)     c           P3            3                   3      b\n"
+            "    0    10 (83.3%)     a        Im-3m         m-3m                 432    red\n"
+            "    1      1 (8.3%)     b         P432          432                 432  green\n"
+            "    2      1 (8.3%)     c           P3            3                   3   blue\n"
             "Properties: iq\n"
             "Scan unit: nm"
         )